+++ ckanext/berlin_dataset_schema/plugin.py
@@ -151,6 +151,14 @@ class Berlin_Dataset_SchemaPlugin(plugins.SingletonPlugin, toolkit.DefaultDatase
                         schema[attribute] = [ required_validator ] + validator_chain
                     else:
                         schema[attribute][0] = required_validator
+
+        resource_schema = schema['resources']
+        for attribute in self.json_schema.resource_schema()['required']:
+            resource_schema[attribute] = [
+                toolkit.get_validator('not_empty'),
+                toolkit.get_validator('unicode_safe')
+            ]
+
         return schema
 
     def _modify_package_schema(self, schema):